Added embed vid iframe template, table of contents template
Added file for foundations for cybersecurity week 1

template gives an iframe template for embedding video. Change the "src" attribute in HTML tag to the destination url.

Template TOC gives table of contents based on the community plugin Dynamic Table of Contents. Set to number style with depth of 6. Adjustable as well.